+++ client/components/FilterBox/FilterBox.jsx
@@ -183,14 +183,16 @@ FilterBox = React.createClass({
         <div className="filter-box ibox">
           <div className="ibox-title">
             <button 
-              onClick={this.props.exportCallback.bind(null,items)} 
+              onClick={this.props.exportCallback?this.props.exportCallback.bind(null,items):null} 
               className="card-button new-card-button pull-right"
               style={{backgroundColor:"transparent",color:"#333"}}
             ><i className="fa fa-download"></i></button>
-            <button 
-              onClick={this.createNewItem} 
-              className="card-button new-card-button pull-right"
-            >+</button>
+            {this.props.newItemCallback==null?null:
+              <button 
+                onClick={this.createNewItem} 
+                className="card-button new-card-button pull-right"
+              >+</button>
+            }
             {title?<h5>{title}</h5>:null}
 
             {!filters?null: