+++ client/views/Dashboard/BarChart/BarChart.jsx
@@ -102,6 +102,7 @@ BarChart = React.createClass({
     	}
 
     	var team = Session.get('selectedTeam');
+        //var team = Teams.findOne({name:"Kaplan Australia Pty Ltd"});
     	if(team) {
     		query["team._id"] = team._id;
     	}